backend/requirements.txt
fastapi==0.128.0
sqlmodel==0.0.31
alembic==1.18.1
uvicorn[standard]==0.40.0
python-multipart==0.0.21
psycopg[binary]==3.3.2
asyncpg==0.30.0
pydantic-settings==2.7.0
structlog==25.5.0
python-dotenv==1.2.1
pyjwt==2.10.1
passlib[argon2]==1.7.4
better-exceptions==0.3.3
pytest==9.0.2
pytest-cov==6.0.0
httpx==0.28.1
slowapi==0.1.9
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2

# MCP and OpenAI Agents SDK (Phase 3)
mcp>=1.2.0
openai-agents==0.7.0
openai-chatkit<=1.4.0
python-multipart==0.0.21